finalizing v0.4

+++ src/main.rs
@@ -145,7 +145,7 @@ fn main() {
     let mut in_selection_mode = false;
     let mut selection_start = 0;
     let mut cmd_history = Vec::<String>::new();
-    let (mut last_mouse_col, mut last_mouse_row) = (0 as u16, 0 as u16);
+    let (mut last_mouse_col, mut last_mouse_row) = (0, 0);
     let mut last_click_time = Instant::now();
 
     //main program loop
